Move color helpers to utils lib
