+++ libs/designer-ui/src/lib/html/plugins/toolbar/DropdownColorPicker.tsx
@@ -3,9 +3,8 @@ import { DropDown } from './helper/Dropdown';
 import { MoveWrapper } from './helper/MoveWrapper';
 import { basicColors, COLORPICKER_HEIGHT as HEIGHT, COLORPICKER_WIDTH as WIDTH } from './helper/constants';
 import type { Position } from './helper/util';
-import { transformColor } from './helper/util';
 import { Text } from '@fluentui/react';
-import { capitalizeFirstLetter } from '@microsoft/utils-logic-apps';
+import { capitalizeFirstLetter, transformColor } from '@microsoft/utils-logic-apps';
 import type { LexicalEditor } from 'lexical';
 import type { ReactNode } from 'react';
 import { useEffect, useMemo, useRef, useState } from 'react';
